refactor: 통계 클래스에서 등수 확인 기능 클래스 분리
