+++ src/test/java/lotto/temp/StaticsTest.java
@@ -38,30 +38,6 @@ class StaticsTest {
     }
 
     @Test
-    @DisplayName("일치하는 개수에 따라 올바른 등수를 나타내는지 확인하는 테스트")
-    void confirmLottoRank(){
-        assertThat(statics.determineRank(2,true)).isEqualTo("꽝");
-        assertThat(statics.determineRank(3,true)).isEqualTo("5등");
-        assertThat(statics.determineRank(4,true)).isEqualTo("4등");
-        assertThat(statics.determineRank(5,false)).isEqualTo("3등");
-        assertThat(statics.determineRank(5,true)).isEqualTo("2등");
-        assertThat(statics.determineRank(6,true)).isEqualTo("1등");
-    }
-
-    @Test
-    @DisplayName("등수 리스트를 받아서 당첨 개수를 제대로 나타내는지 확인하는 테스트")
-    void checkRankingCount(){
-        assertThat(statics.finalRank(List.of("5등","5등","5등","5등","꽝")))
-                .containsEntry("5등",4)
-                .containsEntry("꽝",1);
-        assertThat(statics.finalRank(List.of("5등","4등","3등","5등","꽝")))
-                .containsEntry("5등",2)
-                .containsEntry("4등",1)
-                .containsEntry("3등",1)
-                .containsEntry("꽝",1);
-    }
-
-    @Test
     @DisplayName("수익률 계산을 제대로 하는지 확인하는 테스트")
     void calculateProfit(){
         assertThat(statics.calculateProfit(5000, 5000)).isEqualTo(100);
